move Evaluate related things to the same file, and fixes in analytics
