+++ iOS/Sources/Beagle/Sources/ContextExpression/Expression/Expression.swift
@@ -64,7 +64,7 @@ public extension Expression {
     func evaluate(with view: UIView?) -> T? {
         switch self {
         case let .expression(expression):
-            return view?.evaluate(for: expression).transform()
+            return view?.evaluateExpression(expression).transform()
         case let .value(value):
             return value
         }
