+++ iOS/Sources/Beagle/Sources/ContextExpression/UIView+Context.swift
@@ -76,26 +76,13 @@ extension UIView {
             configBinding(for: expression, completion: completion)
         }
     }
-
-    func evaluate<T: Decodable>(expression: Expression<T>) -> T? {
-        return expression.evaluate(with: self)
-    }
-    
-    func evaluate(for expression: ContextExpression) -> DynamicObject {
-        switch expression {
-        case let .single(expression):
-            return evaluate(for: expression)
-        case let .multiple(expression):
-            return evaluate(for: expression)
-        }
-    }
     
     // MARK: Single Expression
     
     private func configBinding<T: Decodable>(_ binding: Binding, in expression: SingleExpression, completion: @escaping (T?) -> Void) {
         guard let context = getContext(with: binding.context) else { return }
         let closure: (Context) -> Void = { [weak self] context in
-            completion(self?.evaluate(for: expression).transform())
+            completion(self?.evaluateSingle(expression).transform())
         }
         let contextObserver = ContextObserver(onContextChange: closure)
         context.addObserver(contextObserver)
@@ -122,18 +109,7 @@ extension UIView {
         case let .operation(operation):
             configBinding(operation, in: expression, completion: completion)
         }
-        completion(evaluate(for: expression).transform())
-    }
-    
-    private func evaluate(for expression: SingleExpression) -> DynamicObject {
-        switch expression {
-        case let .value(.binding(binding)):
-            return binding.evaluate(in: self)
-        case let .value(.literal(literal)):
-            return literal.evaluate()
-        case let .operation(operation):
-            return operation.evaluate(in: self)
-        }
+        completion(evaluateSingle(expression).transform())
     }
     
     // MARK: Multiple Expression
@@ -141,7 +117,7 @@ extension UIView {
     private func configBinding<T: Decodable>(_ binding: Binding, in expression: MultipleExpression, completion: @escaping (T?) -> Void) {
         guard let context = getContext(with: binding.context) else { return }
         let closure: (Context) -> Void = { [weak self] _ in
-            let value: T? = self?.evaluate(for: expression, contextId: binding.context).transform()
+            let value: T? = self?.evaluateMultiple(expression, contextId: binding.context).transform()
             completion(value)
         }
         let contextObserver = ContextObserver(onContextChange: closure)
@@ -172,21 +148,7 @@ extension UIView {
                 }
             }
         }
-        completion(evaluate(for: expression).transform())
-    }
-    
-    private func evaluate(for expression: MultipleExpression, contextId: String? = nil) -> DynamicObject {
-        var result: String = ""
-        expression.nodes.forEach {
-            switch $0 {
-            case let .expression(expression):
-                let evaluated: String? = evaluateWithCache(for: expression, contextId: contextId).transform()
-                result += evaluated ?? ""
-            case let .string(string):
-                result += string
-            }
-        }
-        return .string(result)
+        completion(evaluateMultiple(expression).transform())
     }
     
     // MARK: Get/Set Context
@@ -222,24 +184,6 @@ extension UIView {
             return global.context.value.value
         }
         return contextMap[contextId]?.value.value
-    }
-    
-    // MARK: Private
-    
-    /// expression last value cache is used only for multiple expressions binding
-    private func evaluateWithCache(for expression: SingleExpression, contextId: String? = nil) -> DynamicObject {
-        switch expression {
-        case let .value(.binding(binding)):
-            if contextId == nil || contextId == binding.context {
-                return evaluate(for: expression)
-            } else {
-                return expressionLastValueMap[binding.rawValue, default: .empty]
-            }
-        case let .value(.literal(literal)):
-            return literal.evaluate()
-        case let .operation(operation):
-            return operation.evaluate(in: self)
-        }
     }
 }
